# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "GPSDEV API",
    "description": "Описание взаимодействия с API для формирования отчетов о перемещениях сотрудников и посещениях ими подопечных.",
    "version": "1.0."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5000/api",
      "description": "Development server"
    },
    {
      "url": "http://127.0.0.1:8000/api",
      "description": "Development server with gunicorn"
    }
  ],
  "paths": {
...
    "/serves": {
      "post": {
        "tags": [
          "Serves"
        ],
        "summary": "Добавить служебные записки",
        "description": "## Добавление новых служебных записок\n| Параметр | Описание |\n| -------- | -------- |\n| `name_id`| name_id сотрудника, по которому добавляется служебная записка |\n| `object_id` | object_id подопечного |\n| `date` | Дата служебной записки |\n| `comment` | Комментарий к записке |\n| `approval` | Статус служебной записки (1 - подтверждено, 3 - на проверку). Необязательный параметр, при его отсутствии по умолчанию проставится 3. Статус 1 разрешено отправлять только для руководителя или администратора. |\n",
        "requestBody": {
          "description": "A JSON User object",
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Serves"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Employee created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Serves"
        ],
        "summary": "Подтверждение служебных записок",
        "description": "## Подтверждение служебных записок\nДля подтверждения служебных записок требуется указать минимальные параметры, чтобы найти служебную записку в базе данных. Единственный доступный параметр для изменения - `approval`.\n  | Параметр | Описание |\n  | -------- | -------- |\n  | `date` | Дата служебной записки |\n  | `name_id`| name_id сотрудника, по которому добавляется служебная записка |\n  | `object_id` | object_id подопечного |\n  | `approval` | Статус служебной записки (1 - подтверждено, 3 - на проверку). Подтвердить служебку может только администратор или руководитель. |\n",
        "requestBody": {
          "description": "A JSON User object",
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "properties": {
                    "date": {
                      "type": "string",
                      "format": "date"
                    },
                    "object_id": {
                      "type": "integer"
                    },
                    "name_id": {
                      "type": "integer"
                    },
                    "approval": {
                      "type": "integer",
                      "default": 3,
                      "enum": [
                        1,
                        3
                      ]
                    }
                  },
                  "required": [
                    "date",
                    "name_id",
                    "object_id",
                    "approval"
                  ]
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/serves/get": {
      "post": {
        "tags": [
          "Serves"
        ],
        "summary": "Запрос служебных записок",
        "description": "Позволяет получить служебные записки со всеми параметрами (комментарий, адрес, статус), но для их запроса потребуется минимальный набор параметров: `name_id`, `object_id`, `date`.\n",
        "requestBody": {
          "description": "Список служебных записок, которые нужно получить",
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "properties": {
                    "object_id": {
                      "type": "integer"
                    },
                    "name_id": {
                      "type": "integer"
                    },
                    "date": {
                      "type": "string",
                      "format": "date"
                    }
                  },
                  "required": [
                    "date",
                    "name_id",
                    "object_id"
                  ]
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "A JSON array of employees",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Serves"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotAllowedError"
                }
              }
            }
          }
        }
      }
    },
    "/serves/delete": {
      "post": {
        "tags": [
          "Serves"
        ],
        "summary": "Удаление служебных записок",
        "description": "Удаление списка служебных записок. Необходимо передать `object_id`, `name_id` и `date` по каждой служебной записке.",
        "requestBody": {
          "description": "A JSON User object",
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "properties": {
                    "date": {
                      "type": "string",
                      "format": "date"
                    },
                    "object_id": {
                      "type": "integer"
                    },
                    "name_id": {
                      "type": "integer"
                    }
                  },
                  "required": [
                    "date",
                    "name_id",
                    "object_id"
                  ]
                }
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/serves/with-coordinates": {
      "post": {
        "tags": [
          "Serves"
        ],
        "summary": "Служебная записка с проверкой нахождения по адресу",
        "description": "## Проверка нахождения по адресу перед добавлением служебки\nНекоторые служебные записки подразумевают нахождение сотрудника по делам подопечного по определенному адресу. В таком случае происходит автоматическое подтверждение нахождения сотрудника по адресу, иначе служебную записку добавить не получится.\n| Параметр | Описание |\n| -------- | -------- |\n| `name_id`| name_id сотрудника, по которому добавляется служебная записка |\n| `object_id` | object_id подопечного |\n| `division` | Номер подразделения, в по которому проверять служебку |\n| `date` | Дата служебной записки |\n| `latitude`, `longitude` | Координаты нахождения сотрудника для проверки |\n| `comment` | Комментарий к записке |\n| `address` |  |\n",
        "requestBody": {
          "description": "A JSON User object",
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ServesWithCoordinates"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Служебная записка была подтверждена и добавлена",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "422": {
            "description": "\"Validation error\" или \"Report error\"",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ReportError"
                    },
                    {
                      "$ref": "#/components/schemas/ValidationError"
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ValidationError": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "422"
          },
          "title": {
            "type": "string",
            "example": "validation error"
          },
          "details": {
            "type": "object",
            "example": {}
          }
        }
      },
...
      "NotAllowedError": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "403"
          },
          "title": {
            "type": "string",
            "example": "not allowed"
          },
          "details": {
            "type": "object",
            "example": {}
          }
        }
      },
      "ReportError": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "422"
          },
          "title": {
            "type": "string",
            "example": "report error"
          },
          "details": {
            "type": "object",
            "example": {}
          }
        }
      },
...
      "Serves": {
        "type": "object",
        "properties": {
          "name_id": {
            "type": "integer"
          },
          "object_id": {
            "type": "integer"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "comment": {
            "type": "string",
            "maxLength": 200
          },
          "approval": {
            "type": "integer",
            "default": 3,
            "enum": [
              1,
              3
            ]
          },
          "id": {
            "type": "integer",
            "readOnly": true
          },
          "object": {
            "readOnly": true,
            "type": "string",
            "maxLength": 100
          },
          "name": {
            "readOnly": true,
            "type": "string"
          }
        },
        "required": [
          "comment",
          "date",
          "name_id",
          "object_id"
        ]
      },
      "ServesWithCoordinates": {
        "type": "object",
        "properties": {
          "name_id": {
            "type": "integer"
          },
          "object_id": {
            "type": "integer"
          },
          "division": {
            "type": "integer"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "longitude": {
            "type": "number"
          },
          "latitude": {
            "type": "number"
          },
          "comment": {
            "type": "string",
            "maxLength": 200
          },
          "address": {
            "type": "string",
            "maxLength": 200
          }
        },
        "required": [
          "comment",
          "date",
          "division",
          "latitude",
          "longitude",
          "name_id",
          "object_id"
        ]
      },
...
    },
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "JSESSIONID"
      }
    }
  },
  "security": [
    {
      "cookieAuth": []
    }
  ],
...
}